Version 0.04
savage can now dynamically utilize cap2hccapx to verify WPA handshakes.  This method is 100% accurate when compared to aircrack-ng which does struggle in edge cases.  If cap2hccapx is in path, savage will use it automatically and result in *.hccapx files instead of *.cap files in the handshake output dir.  Currently the only way to disable this functionality is by commenting out line 356 in this commit.

cap2hccapx is packaged with hashcat-utils: https://github.com/hashcat/hashcat-utils